게시글 삭제시 로그인상태체크와 권한 유무체크하기/게시글수정 구현중

+++ src/main/java/com/hsn/exam/demo/controller/UsrArticleController.java
@@ -68,11 +68,34 @@ public class UsrArticleController {
 
 		@RequestMapping("/usr/article/doDelete")
 		@ResponseBody
-		public ResultData<Integer> doDelete(int id) {
+		public ResultData<Integer> doDelete(HttpSession httpSession,int id) {
+			
+			boolean isLogined = false;
+			
+			int loginedMemberId  = 0;
+			
+			if(httpSession.getAttribute("loginedMemberId")!=null) {
+				
+				isLogined =true;
+				
+				loginedMemberId = (int)httpSession.getAttribute("loginedMemberId");
+				
+			}
+			
+			if(isLogined==false) {
+				return ResultData.from("F-3", "로그인후 이용해주시기 바랍니다.");
+			}
+			
+			
 			Article article = articleService.getArticle(id);
 
 			if (article == null) {
 				return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다", id), id);
+			}
+			
+			if(article.getMemberId()!=loginedMemberId) {
+				
+				return ResultData.from("F-2", Ut.f("%d번 게시물에 대한 권한이 없습니다.", id));
 			}
 
 			articleService.deleteArticle(id);
@@ -82,11 +105,33 @@ public class UsrArticleController {
 
 		@RequestMapping("/usr/article/doModify")
 		@ResponseBody
-		public ResultData<Integer> doModify(int id, String title, String body) {
+		public ResultData<Integer> doModify(HttpSession httpSession,int id, String title, String body) {
+			
+			boolean isLogined = false;
+			
+			int loginedMemberId  = 0;
+			
+			if(httpSession.getAttribute("loginedMemberId")!=null) {
+				
+				isLogined =true;
+				
+				loginedMemberId = (int)httpSession.getAttribute("loginedMemberId");
+				
+			}
+			
+			if(isLogined==false) {
+				return ResultData.from("F-3", "로그인후 이용해주시기 바랍니다.");
+			}
+			
 			Article article = articleService.getArticle(id);
 
 			if (article == null) {
 				return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다", id), id);
+			}
+			
+			if(article.getMemberId()!=loginedMemberId) {
+				
+				return ResultData.from("F-2", Ut.f("%d번 게시물에 대한 수정권한이 없습니다.", id));
 			}
 
 			articleService.modifyArticle(id, title, body);
